+++ include/H2DE/objects/H2DE_button_object.h
@@ -12,8 +12,8 @@ private:
     H2DE_ButtonObject(H2DE_Engine* engine, const H2DE_ObjectData& objectData, const H2DE_ButtonObjectData& buttonObjectData);
     ~H2DE_ButtonObject() override;
 
-    void updateSurfaceBuffers() override;
-    void updateMaxRadius() override;
+    void refreshSurfaceBuffers() override;
+    void refreshMaxRadius() override;
 
 public:
     inline H2DE_Texture* addTexture(const std::string& name, const H2DE_SurfaceData& surfaceData, const H2DE_TextureData& textureData) {