+++ src/surfaces/H2DE_color.cpp
@@ -11,6 +11,14 @@ H2DE_Color::~H2DE_Color() {
 
 }
 
+// GETTER
+bool H2DE_Color::isVisible() const {
+    bool surfaceIsNotHidden = !(isHidden());
+    bool colorIsVisible = (colorData.color.isVisible());
+    
+    return (surfaceIsNotHidden && colorIsVisible);
+}
+
 // SETTER
 unsigned int H2DE_Color::setColor(const H2DE_ColorRGB& color, unsigned int duration, H2DE_Easing easing, const std::function<void()>& completed, bool pauseSensitive) {
     return H2DE_LerpManager::lerp(engine, colorData.color, color, duration, easing, [this](H2DE_ColorRGB iv) {