+++ src/surfaces/H2DE_sprite.cpp
@@ -1,4 +1,5 @@
 #include "H2DE/surfaces/H2DE_sprite.h"
+#include "H2DE/engine/H2DE_lerp_manager.h"
 
 // INIT
 H2DE_Sprite::H2DE_Sprite(H2DE_Engine* e, H2DE_Object* o, const H2DE_SurfaceData& sd, const H2DE_SpriteData& td) : H2DE_Surface(e, o, sd), spriteData(td) {
@@ -40,4 +41,61 @@ std::optional<H2DE_PixelRect> H2DE_Sprite::getSrcRect() const {
     srcRect.addScale(spriteData.size);
 
     return srcRect;
+}
+
+bool H2DE_Sprite::isVisible() const {
+    bool surfaceIsNotHidden = !(isHidden());
+    bool colorIsVisible = (spriteData.color.isVisible());
+    
+    return (surfaceIsNotHidden && colorIsVisible);
+}
+
+// SETTER
+
+// -- no lerp
+void H2DE_Sprite::setTextureName(const std::string& textureName) {
+    spriteData.textureName = textureName;
+    object->refreshSurfaceBuffers();
+}
+
+void H2DE_Sprite::setColor(const H2DE_ColorRGB& color) {
+    spriteData.color = color;
+    object->refreshSurfaceBuffers();
+}
+
+void H2DE_Sprite::setStartingPos(const H2DE_PixelPos& startingPos) {
+    spriteData.startingPos = startingPos;
+    object->refreshSurfaceBuffers();
+}
+
+void H2DE_Sprite::setSize(const H2DE_PixelSize& size) {
+    spriteData.size = size;
+    object->refreshSurfaceBuffers();
+}
+
+void H2DE_Sprite::setSpacing(int spacing) {
+    spriteData.spacing = spacing;
+    object->refreshSurfaceBuffers();
+}
+
+void H2DE_Sprite::setNbFrame(unsigned int nbFrame) {
+    spriteData.nbFrame = nbFrame;
+    object->refreshSurfaceBuffers();
+}
+
+void H2DE_Sprite::setDelay(unsigned int delay) {
+    spriteData.delay = delay;
+    object->refreshSurfaceBuffers();
+}
+
+void H2DE_Sprite::setPauseSensitive(bool pauseSensitive) {
+    spriteData.pauseSensitive = pauseSensitive;
+    object->refreshSurfaceBuffers();
+}
+
+// -- lerp
+unsigned int H2DE_Sprite::setColor(const H2DE_ColorRGB& color, unsigned int duration, H2DE_Easing easing, const std::function<void()>& completed, bool pauseSensitive) {
+    return H2DE_LerpManager::lerp(engine, spriteData.color, color, duration, easing, [this](H2DE_ColorRGB iv) {
+        setColor(iv);
+    }, completed, pauseSensitive);
 }